chore(repo): split node/web helpers to avoid externalizing path/fs
