+++ packages/netlogo-docs/src/helpers-node/nuxt.ts
