+++ packages/netlogo-docs/src/helpers/index.ts
@@ -1,4 +1,1 @@
-export * from "./navigation";
-export * from "./nuxt";
-export * from "./routes";
 export * from "./slugify";
